Move alloy_rpc_types_eth::TransactionInfo into crate alloy-network-primitives
