+++ crates/network-primitives/src/lib.rs
@@ -6,6 +6,9 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+mod transaction;
+pub use transaction::TransactionInfo;
+
 mod traits;
 pub use traits::{BlockResponse, HeaderResponse, ReceiptResponse, TransactionResponse};
